Reconstruct equipment comparison displays and panel motion

+++ include/game/menu_number.h
@@ -6,7 +6,7 @@ extern "C" {
 #endif
 /* Build a 32-by-16-pixel OBJ strip from transparent 4bpp glyphs. */
 void MenuNumber_DrawUnavailable(int tile, u32 x);
-void MenuNumber_DrawValue(int tile, u32 x, u16 value, int wide);
+void MenuNumber_DrawValue(int tile, u32 x, int value, int wide);
 void MenuNumber_PrepareGlyphs(void);
 #ifdef __cplusplus
 }
